requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy
matplotlib
scikit-learn
tqdm
pillow
opencv-python
seaborn